pyproject.toml
[project]
name = "repl-nix-mmac-automaticmouseclicker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cairosvg>=2.7.1",
    "keyboard>=0.13.5",
    "pillow>=11.0.0",
    "pyautogui>=0.9.54",
    "pyinstaller>=6.11.0",
    "requests>=2.32.3",
]
